Added an ok in the end of the tests that use a helper function with asserts.
A failing test, run by foreach, does not show up in the stack trace if
it fails in it's last statement and that statement is a call to
another function where the asserts are.  This is typically the case
for all the called_ tests.  This can be helped by putting an ok in
the end.